Fix more MSVC warnings

+++ include/dkm.hpp
@@ -142,7 +142,7 @@ uint32_t closest_mean(const std::array<T, N>& point, const std::vector<std::arra
 			index = i;
 		}
 	}
-	return index;
+	return static_cast<uint32_t>(index);
 }
 
 /*
